Multiplatform fix to 2 race conditions
fcntl is only useful when you want to edit a file safely, if you need to
ensure your program is running only once we ought to use atomic os.open
flags. That way race condition between opening of file and locking does
not happen. Then closing the file descriptor and unlinking is enough to
cancel locking as well as perform it atomically eleminating race
condition when for example the script runs too fast and reaches GC
before the other instance accidentally runs and it hits the small time
space between unlock in fcntl and removal of the lockfile we will have
race condition. Though it is not that important because most it can do
is break our script exiting routine. Which already should account for
that.