Fix with proper handling of already running process
Got a nice function from stackoverflow for checking whether pid is on.
Rewrote all code to low-level os module calls on file descriptor.
Removed not needed vars and checks.

Code is not tested BEWARE! Should be tested because some optimizations
are here to remove any possibility of race conditions, however I did not
test the exceptions and whether the code is clean yet.