Fixing platforms
Windows and Linux do not support true atomic locking upon opening which
does not make it easier. I try to overcame this developers' shortcomings
by trying to get an exclusive lock after reading the file. Basically if
there  is no lock on file it should be closed. But I'm trying to do that
a bit safer.